+++ changelog.md
@@ -5,6 +5,7 @@ Bumping the minor version of openbci-utilities allows this module to move forwar
 ### Breaking Changes
 
 * Drop support for node 4 because of new OpenBCI utilities
+* Connect will now automatically call `softReset` and resolve once it's complete.
 
 # 1.0.8